test(kb): refactor rollback compensation tests
Move cleanup_decision test from test_pipeline_compatibility.py
to test_api_compatibility.py where it belongs. Parameterize
two per-boundary compensation tests into one using pytest
parametrize (all_succeed / document_fails cases). Remove
duplicate assertions and unused imports.
